Moved sample event data to /src/lib/server/events/data/
